Rework DefoldSdkService to deal with concurrent download/acquiring sdk for building.

+++ server/src/main/java/com/defold/extender/services/DefoldSdkService.java
@@ -4,6 +4,8 @@ import com.defold.extender.ExtenderException;
 import com.defold.extender.ZipUtils;
 import com.defold.extender.log.Markers;
 import com.defold.extender.metrics.MetricsWriter;
+import com.defold.extender.services.data.DefoldSdk;
+
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,6 +23,7 @@ import jakarta.annotation.PreDestroy;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
@@ -28,6 +31,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.Comparator;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutionException;
 
 @Service
 public class DefoldSdkService {
@@ -60,8 +66,8 @@ public class DefoldSdkService {
     private final boolean cacheClearOnExit;
 
     private final MeterRegistry meterRegistry;
-    // private final Counter counterSdkGet;
-    // private final Counter counterSdkGetDownload;
+    private final ConcurrentHashMap<String, CompletableFuture<DefoldSdk>> operationCache = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<String, Integer> cacheReferenceCount;
 
     DefoldSdkService(@Value("${extender.sdk.location}") String baseSdkDirectory,
                      @Value("${extender.sdk.cache-size}") int cacheSize,
@@ -71,33 +77,14 @@ public class DefoldSdkService {
         this.cacheSize = cacheSize;
         this.meterRegistry = meterRegistry;
         this.cacheClearOnExit = cacheClearOnExit;
-        // this.counterSdkGet = gaugeService.counter("counter.service.sdk.get");
-        // this.counterSdkGetDownload = gaugeService.counter("counter.service.sdk.get.download");
 
         this.dynamoHome = System.getenv("DYNAMO_HOME") != null ? new File(System.getenv("DYNAMO_HOME")) : null;
+        this.cacheReferenceCount = new ConcurrentHashMap<>(this.cacheSize + 10);
 
         LOGGER.info("SDK service using directory {} with cache size {}", baseSdkDirectory, cacheSize);
 
         if (!Files.exists(this.baseSdkDirectory)) {
             Files.createDirectories(this.baseSdkDirectory);
-        }
-    }
-
-    // Helper function to move the SDK directories
-    public static void Move(Path source, Path target) {
-        try {
-            Files.move(source, target, StandardCopyOption.ATOMIC_MOVE);
-        } catch (IOException e) {
-            // If the target path suddenly exists, and the source path still exists,
-            // then we failed with the atomic move, and we assume another job succeeded with the download
-            if (Files.exists(source) && Files.exists(target)) {
-                LOGGER.info("Defold SDK version {} was downloaded by another job in the meantime", source.toString());
-                try {
-                    FileUtils.deleteDirectory(source.toFile());
-                } catch (IOException e2) {
-                    LOGGER.error(Markers.SERVER_ERROR, "Failed to delete temp sdk directory {}: {}", source.toString(), e2.getMessage());
-                }
-            }
         }
     }
 
@@ -109,26 +96,45 @@ public class DefoldSdkService {
         return sdkVersion == null || LOCAL_VERSION.equals(sdkVersion) || System.getenv("DYNAMO_HOME") != null;
     }
 
-    public File getSdk(String hash) throws IOException, URISyntaxException, ExtenderException {
-        return isLocalSdk(hash) ? getLocalSdk() : getRemoteSdk(hash);
+    public DefoldSdk getSdk(String hash) throws ExtenderException {
+        if (isLocalSdk(hash)){
+            return getLocalSdk();
+        }
+        // use ConcurrentHashMap with CompletableFuture here to avoid situation when
+        // several builds needs the same defoldsdk which doesn't exist locally. So one build job starts downloading,
+        // all other jobs wait for download complete and all of then continue running.
+        CompletableFuture<DefoldSdk> operation = operationCache.computeIfAbsent(hash, this::getRemoteSdk);
+        try (DefoldSdk sdk = operation.get()){
+            if (sdk == null) {
+                throw new ExtenderException(String.format("The given sdk does not exist: %s", hash));
+            }
+            if (!sdk.isValid()) {
+                throw new ExtenderException(String.format("The given sdk still does not exist: %s", hash));
+            }
+            evictCache();
+            LOGGER.info("Using Defold SDK version {}", hash);
+            return DefoldSdk.copyOf(sdk);
+        } catch (InterruptedException | ExecutionException e) {
+            LOGGER.error(String.format("The given sdk cannot be downloaded: %s", hash), e);
+            throw new ExtenderException(String.format("The given sdk cannot be downloaded: %s", hash));
+        } finally {
+            operationCache.remove(hash);
+        }
     }
 
-    public File getRemoteSdk(String hash) throws IOException, URISyntaxException, ExtenderException {
-        long methodStart = System.currentTimeMillis();
-
-        // Define SDK directory for this version
-        File sdkDirectory = new File(baseSdkDirectory.toFile(), hash);
-
-        // If directory does not exist, create it and download SDK
-        if (!Files.exists(sdkDirectory.toPath())) {
-            File lockFile = new File(baseSdkDirectory.toFile(), "tmp" + hash + ".lock");
-
-            if (lockFile.createNewFile()) { // atomic creation of lock file
-                try {
-                    boolean sdkFound = false;
-                    for (String url_pattern : REMOTE_SDK_URL_PATTERNS) {
-
-                        URL url = new URL(String.format(url_pattern, hash));
+    public CompletableFuture<DefoldSdk> getRemoteSdk(String hash) {
+        return CompletableFuture.supplyAsync(() -> {
+            long methodStart = System.currentTimeMillis();
+            // Define SDK directory for this version
+            File sdkDirectory = new File(baseSdkDirectory.toFile(), hash);
+            File sdkRootDirectory = new File(sdkDirectory, "defoldsdk");
+            DefoldSdk sdk = new DefoldSdk(sdkRootDirectory, hash, this);
+            // If directory does not exist, create it and download SDK
+            if (!Files.exists(sdkDirectory.toPath())) {
+                boolean sdkFound = false;
+                for (String url_pattern : REMOTE_SDK_URL_PATTERNS) {
+                    try {
+                        URL url = URI.create(String.format(url_pattern, hash)).toURL();
 
                         ClientHttpRequestFactory clientHttpRequestFactory = new SimpleClientHttpRequestFactory();
                         ClientHttpRequest request = clientHttpRequestFactory.createRequest(url.toURI(), HttpMethod.GET);
@@ -148,73 +154,67 @@ public class DefoldSdkService {
                             InputStream body = response.getBody();
                             ZipUtils.unzip(body, tmpSdkDirectory.toPath());
 
-                            Move(tmpSdkDirectory.toPath(), sdkDirectory.toPath());
-
+                            Files.move(tmpSdkDirectory.toPath(), sdkDirectory.toPath(), StandardCopyOption.ATOMIC_MOVE);
                             sdkFound = true;
                             break;
                         }
-                    }
-
-                    if (!sdkFound) {
-                        throw new ExtenderException(String.format("The given sdk does not exist: %s", hash));
-                    }
-
-                    // Delete old SDK:s
-                    Comparator<Path> lastModifiedComparator = Comparator.comparing(path -> path.toFile().lastModified());
-
-                    Files
-                            .list(baseSdkDirectory)
-                            .filter(path -> !path.getFileName().toString().startsWith("tmp"))
-                            .sorted(lastModifiedComparator.reversed())
-                            .skip(cacheSize)
-                            .forEach(this::deleteCachedSdk);
-
-                    //counterSdkGetDownload.increment();
-                    MetricsWriter.metricsCounterIncrement(meterRegistry, "extender.service.sdk.get.download", "sdk", hash);
-                } finally {
-                    lockFile.delete();
-                }
-            } else {
-                LOGGER.info("Waiting for Defold SDK version {} to download ...", hash);
-                // We have to wait for the lock file to disappear
-                int seconds = 120; // Downloading an sdk takes ~30-50 seconds
-                while (Files.exists(lockFile.toPath()) && seconds > 0) {
-                    int step = 2;
-                    seconds -= step;
-                    try {
-                        Thread.sleep(step * 1000);
-                    } catch (InterruptedException e) {
-                        // pass
+                    } catch (IOException|URISyntaxException exc) {
+                        LOGGER.error("Error downloading defoldsdk", exc);
                     }
                 }
 
-                // Now check again that the SDK was downloaded
-                if (!Files.exists(sdkDirectory.toPath())) {
-                    throw new ExtenderException(String.format("The given sdk still does not exist: %s", hash));
+                if (!sdkFound) {
+                    sdk.close();
+                    return null;
                 }
+
+                MetricsWriter.metricsCounterIncrement(meterRegistry, "extender.service.sdk.get.download", "sdk", hash);
             }
-        }
-
-        LOGGER.info("Using Defold SDK version {}", hash);
-
-        MetricsWriter.metricsTimer(meterRegistry, "extender.service.sdk.get.duration", System.currentTimeMillis() - methodStart, "sdk", hash);
-
-        return new File(sdkDirectory, "defoldsdk");
+            MetricsWriter.metricsTimer(meterRegistry, "extender.service.sdk.get.duration", System.currentTimeMillis() - methodStart, "sdk", hash);
+            return sdk;
+        });
     }
 
-    public File getLocalSdk() {
+    public DefoldSdk getLocalSdk() {
         LOGGER.info("Using local Defold SDK at {}", dynamoHome.toString());
-        return dynamoHome;
+        return new DefoldSdk(dynamoHome, LOCAL_VERSION, this);
     }
 
     public boolean isLocalSdkSupported() {
         return dynamoHome != null;
     }
 
+    protected void evictCache() {
+        synchronized (cacheReferenceCount) {
+            try {
+                LOGGER.info("Cache eviction called");
+                // Delete old SDK:s
+                Comparator<Path> refCountComparator = Comparator.comparing(path -> getSdkRefCount(path.getFileName().toString()));
+
+                Files
+                        .list(baseSdkDirectory)
+                        .filter(path -> !path.getFileName().toString().startsWith("tmp")
+                                    && !path.toString().endsWith(".delete")
+                                    && !path.getFileName().toString().equals(TEST_SDK_DIRECTORY))
+                        .sorted(refCountComparator.reversed())
+                        .skip(cacheSize)
+                        .forEach(this::deleteCachedSdk);
+            } catch (IOException exc) {
+                LOGGER.error("Error during cache eviction", exc);
+            }
+        }
+    }
+
     private void deleteCachedSdk(Path path) {
+        String sdkHash = path.getFileName().toString();
+        if (getSdkRefCount(sdkHash) != 0) {
+            LOGGER.warn(String.format("Sdk %s remove skipped due to non-zero ref count", sdkHash));
+            return;
+        }
         try {
+            LOGGER.info(String.format("Cleanup sdk %s", path));
             File tmpDir = new File(path.toString() + ".delete");
-            Move(path, tmpDir.toPath());
+            Files.move(path, tmpDir.toPath(), StandardCopyOption.ATOMIC_MOVE);
             FileUtils.deleteDirectory(tmpDir);
         } catch (IOException e) {
             LOGGER.error(Markers.CACHE_ERROR, "Failed to delete cached SDK at " + path.toAbsolutePath().toString(), e);
@@ -222,7 +222,6 @@ public class DefoldSdkService {
     }
 
     @PreDestroy
-    @SuppressWarnings("unused")
     public void destroy() {
         if (!this.cacheClearOnExit) {
             LOGGER.info("Skipping cleanup of SDK cache");
@@ -239,12 +238,12 @@ public class DefoldSdkService {
     }
 
     private String getLocalPlatformSdkMappings(String hash) throws IOException {
-        return new String(Files.readAllBytes(Path.of(getLocalSdk().getAbsolutePath(), "platform.sdks.json")), StandardCharsets.UTF_8);
+        return new String(Files.readAllBytes(Path.of(getLocalSdk().toFile().getAbsolutePath(), "platform.sdks.json")), StandardCharsets.UTF_8);
     }
 
     private String getRemotePlatformSdkMappings(String hash) throws IOException, URISyntaxException, ExtenderException {
         for (String url_pattern : REMOTE_MAPPINGS_URL_PATTERNS) {
-            URL url = new URL(String.format(url_pattern, hash));
+            URL url = URI.create(String.format(url_pattern, hash)).toURL();
 
             ClientHttpRequestFactory clientHttpRequestFactory = new SimpleClientHttpRequestFactory();
             ClientHttpRequest request = clientHttpRequestFactory.createRequest(url.toURI(), HttpMethod.GET);
@@ -266,5 +265,19 @@ public class DefoldSdkService {
 
     public String getPlatformSdkMappings(String hash) throws IOException, URISyntaxException, ExtenderException {
         return isLocalSdk(hash) ? getLocalPlatformSdkMappings(hash) : getRemotePlatformSdkMappings(hash);
+    }
+
+    public void acquireSdk(String hash) {
+        LOGGER.info(String.format("Acquire sdk %s", hash));
+        cacheReferenceCount.compute(hash, ( key, value) -> value == null ? 1 : value + 1);
+    }
+
+    public void releaseSdk(String hash) {
+        LOGGER.info(String.format("Release sdk %s", hash));
+        cacheReferenceCount.compute(hash, (key, value) -> value - 1);
+    }
+
+    public Integer getSdkRefCount(String hash) {
+        return cacheReferenceCount.getOrDefault(hash, 0);
     }
 }
